+++ discopy/quantum/channel.py
@@ -51,7 +51,7 @@ from discopy.quantum.circuit import (
     Digit, Qudit)
 from discopy.quantum.gates import Discard, Measure, MixedState, Encode, Scalar
 from discopy.tensor import Dim, Tensor
-from discopy.utils import assert_isinstance
+from discopy.utils import assert_isinstance, unbiased
 
 
 class CQ:
@@ -174,9 +174,9 @@ class Channel(Tensor):
         assert_isinstance(dom, CQ)
         return cls(Tensor[cls.dtype].id(dom.to_dim()).array, dom, dom)
 
-    def then(self, other: Channel = None, *others: Channel) -> Channel:
-        if other is None or others:
-            return super().then(other, *others)
+    @unbiased
+    def then(self, other: Channel) -> Channel:
+        """ The composition of two channels, through their tensors. """
         assert_isinstance(other, type(self))
         array = (self.to_tensor() >> other.to_tensor()).array
         return type(self)(array, self.dom, other.cod)